Removed all panels and moved them into new 'Menu' system at bottom of the screen.
