--- webstack/apps/webapp/src/app/pages/board/layers/ui/components/Panels/index.ts
@@ -1,16 +1,0 @@
-/**
- * Copyright (c) SAGE3 Development Team 2022. All Rights Reserved
- * University of Hawaii, University of Illinois Chicago, Virginia Tech
- *
- * Distributed under the terms of the SAGE3 License.  The full license is in
- * the file LICENSE, distributed as part of this software.
- */
-
-export * from './Controller/Controller';
-export * from './Applications/ApplicationsPanel';
-export * from './Asset/AssetsPanel';
-export * from './Navigation/NavigationPanel';
-export * from './Users/UsersPanel';
-export * from './Annotations/AnnotationsPanel';
-export * from './Plugins/PluginsPanels';
-export * from './Kernels/KernelsPanel';
