+++ webstack/apps/webapp/src/app/pages/board/layers/ui/components/Toolbar/index.ts
@@ -1,0 +1,2 @@
+export * from './ToolbarButton';
+export * from './ToolbarMenus';
