+++ webstack/apps/webapp/src/app/pages/board/layers/ui/components/index.ts
@@ -10,7 +10,6 @@ export * from './AppToolbar';
 export * from './BoardContextMenu';
 export * from './ClearBoardModal';
 export * from './LassoToolbar';
-export * from './Panels';
 export * from './Twilio';
 export * from './PresenceFollow';
 export * from './BoardTitle';
@@ -18,3 +17,4 @@ export * from './TagsDisplay';
 export * from './Interactionbar';
 export * from './InteractionbarShortcuts';
 export * from './ScreenshareMenu';
+export * from './Toolbar';
