+++ webstack/libs/frontend/src/lib/stores/index.ts
@@ -11,7 +11,6 @@ export * from './user';
 export * from './room';
 export * from './board';
 export * from './asset';
-export * from './panel';
 export * from './presence';
 export * from './twilio';
 export * from './ui';
